requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.9
elasticsearch==8.11.0
bcrypt==4.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
websockets==12.0 